patch
patch

+++ LostArkLogger/Packets/Struct11.cs
@@ -6,13 +6,13 @@ namespace LostArkLogger
     {
         public Struct11(BitReader reader)
         {
-            num = reader.ReadUInt16();
+            num = reader.ReadUInt32();
             for(var i = 0; i < num; i++)
             {
-                field0s.Add(reader.ReadUInt32());
+                field0s.Add(reader.ReadByte());
             }
         }
-        public UInt16 num;
-        public List<UInt32> field0s = new List<UInt32>();
+        public UInt32 num;
+        public List<Byte> field0s = new List<Byte>();
     }
 }
